Create texsetup.md
Process to ensure Windows R Studio is able to KNIT a pdf. I kept getting some fancyvrb.  not available. It finally looked like it was just a package that  I needed to install in MikTeX. While MikTeX should install it on the fly I dont know why it dint work .